requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
Werkzeug==3.0.3
python-dotenv==1.0.1
requests==2.32.3